backend/requirements.txt
Flask==3.1.0
pandas==2.2.3
openai==1.55.3
python-dotenv==1.0.1
flask-cors==5.0.0
scikit-learn==1.5.2